chore: improve build process
Split codegen part from standalone lib functions
