+++ recipes/helper/main.h
@@ -3,52 +3,6 @@
 #ifdef REFLECTC_HELPER
 #ifdef REFLECTC_HEADER
 
-/*#! #include "submodules/oa-hash/oa_hash.h" */
-
-enum reflectc_modes { REFLECTC_MODES_WRITE, REFLECTC_MODES_READONLY };
-
-enum reflectc_types {
-    REFLECTC_TYPES__char,
-    REFLECTC_TYPES__short,
-    REFLECTC_TYPES__int,
-    REFLECTC_TYPES__long,
-    REFLECTC_TYPES__float,
-    REFLECTC_TYPES__double,
-    REFLECTC_TYPES__bool,
-    REFLECTC_TYPES__struct,
-    REFLECTC_TYPES__EXTEND
-};
-
-/* forward definition */
-struct reflectc;
-/**/
-
-typedef struct reflectc *(*const reflectc_from_type_t)(
-                            void *, enum reflectc_modes, struct reflectc *);
-
-struct reflectc {
-    const unsigned ptr_depth;
-    const size_t size;
-    const struct {
-        const size_t len;
-        const char *const buf;
-    } name;
-    const enum reflectc_types type;
-    reflectc_from_type_t init;
-    void *value;
-    struct oa_hash ht;
-    enum reflectc_modes mode;
-};
-
-struct reflectc *reflectc_get(struct reflectc *reflectc,
-                            const char *key,
-                            const size_t key_len);
-
-struct reflectc *reflectc_set(struct reflectc *reflectc,
-                            const char *key,
-                            const size_t key_len,
-                            void *value);
-
 #define REFLECTC_STRUCT_public(_type)                                         \
     struct reflectc *reflectc_from_##_type(struct _type *self,                \
                                          enum reflectc_modes mode,            \
@@ -60,23 +14,7 @@ struct reflectc *reflectc_set(struct reflectc *reflectc,
 
 /*#! #include <stdlib.h> */
 /*#! #include <string.h> */
-
-struct reflectc *
-reflectc_get(struct reflectc *reflectc, const char *key, const size_t key_len)
-{
-    return oa_hash_get(&reflectc->ht, key, key_len);
-}
-
-struct reflectc *
-reflectc_set(struct reflectc *reflectc,
-            const char *key,
-            const size_t key_len,
-            void *value)
-{
-    return (reflectc->mode & REFLECTC_MODES_WRITE)
-               ? oa_hash_set(&reflectc->ht, key, key_len, value)
-               : NULL;
-}
+/*#! #include "reflect-c_internal.h" */
 
 #define REFLECTC_STRUCT_private(_type)                                        \
     static struct reflectc *reflectc_from_##_type(struct _type *self,         \
