+++ reflect-c_RECIPES.PRE.h
@@ -1,7 +1,7 @@
 #ifndef REFLECTC_ENTRY
 #   error "Missing REFLECTC_ENTRY definition"
 #else
-#   include "reflect-c.h"
+#   include "reflect-c_RECIPES_SYMBOLS.h"
 #   include "recipes/definitions/main.h"
 #   include "recipes/helper/main.h"
 #   include "recipes/struct-init/main.h"
